requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
mysql-connector-python==8.3.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
urllib3==2.2.1
